+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+ - Added basic code examples under `Samples~/Basic`.
+
 ### Changed
  - Updated `.gitignore` to ignore more IDE cache directories (`.vscode` and `.idea`).
  - Updated keywords on the `package.json`.